First draft
Non-functional prototype. Filters show up but don't function, mustache templates are overly complex and broken. Lots of back and forth getting the js to work so code is available in multiple places. Will use learnings of that attempt to rebuild a more functional version. Will provide original plugin as context in it;s entirety as welll as better instructions for how I expect the fitlers to function